Proposed grammar
Please review this carefully! This will need to updated following changes to
the language. Grammars make is more clear what is and is not part of the
official language. In my limited experience they make implementing parsers
easier too. The syntax here is based on Extended BNF following conventions in
Python's Grammar.